added a sane plist file for the emoji definitions
note - do not use the csv file directly.  It was created in MS Excel which creates some weird artifacts in the file.  I had to open it in Numbers, export to CSV, then used an online tool (http://www.mindsizzlers.com/case-studies/tools/csv-to-plist-online-conversion-tool/) to convert that CSV to a plist file.  Also in Numbers I had to delete empty columns, and add missing header info.