Added example images for app home and info
Added two quick and hopefully simple examples for app home screen and
info screen.

The app home screen is the default app and provides some basic
functionality for users who download the app initially. It’ll allow
them to search and copy an emoji to the clipboard.

The info button provides a simple tutorial on how to enable the
keyboard so we don’t get 1 star reviews on this. It can also be shown
as a first time experience then dismissed. (optional)